Split out searcher/grep.rs
